+++ tmpl/vue/elements/func.vue
@@ -1,7 +1,6 @@
 <div :id="func.name" class="item border-bottom">
 <h5 class="alert alert-primary">
-  <!--all in one line to prevent creating redundant spaces during vue render-->
-  <span :class="{ deprecated: !!func.deprecated }"><a :href="func.codeLink">{{ func.name }}</a></span>(<template v-if="func.paramsForMethods && func.paramsForMethods.length > 0"><template v-for="param in func.paramsForMethods.slice(0, -1)">{{ param.name }}<span v-if="param.optional" class="badge small-badge badge-light">opt</span><span v-if="param.type && param.type.length > 0">: <span v-for="(pt, idx) in param.type"><template v-html="pt"></template><template v-if="idx<param.type.length-1"> | </template></span></span><span>, </span></template>{{ func.paramsForMethods.slice(-1)[0].name }}<span v-if="func.paramsForMethods.slice(-1)[0].optional" class="badge small-badge badge-light">opt</span><span v-if="func.paramsForMethods.slice(-1)[0].type && func.paramsForMethods.slice(-1)[0].type.length > 0">: <span v-html="func.paramsForMethods.slice(-1)[0].type[0]"></span></span></template>)
+  <span :class="{ deprecated: !!func.deprecated }"><a :href="func.codeLink">{{ func.name }}</a></span>(<func-signature :func="func"></func-signature>)
   <template v-if="func.returns">→
     <template v-for="ret in (func.returns[0].type ? func.returns[0].type.slice(0, -1) : [])"><span
         v-html="ret"></span><span> | </span></template>
@@ -22,34 +21,7 @@
     <p class="arguments-title">Return:</p>
     <template v-html="func.returns[0].description"></template>
   </template>
-  <div v-if="func.paramsForMethods && func.paramsForMethods.length > 0" class="arguments">
-    <p class="arguments-title">Arguments info:</p>
-    <ul>
-      <li v-for="param in func.paramsForMethods">
-        <code>{{ param.name }}</code>
-        <span v-if="param.defaultvalue !== undefined"> = {{ param.defaultvalue }} </span>:
-        <span v-for="(pt, idx) in param.type">
-          <template v-html="pt"></template>
-          <template v-if="idx<param.type.length-1">| </template>
-        </span>
-        <div v-if="param.props">
-          <!--<p>Properties</p>-->
-          <ul>
-            <li v-for="prop in param.props">
-              <code>{{ prop.name }}</code><span v-if="prop.defaultvalue">={{ prop.defaultvalue }}</span>:
-              <span
-                  v-if="prop.type" v-html="prop.type[0]"></span>
-              <!--<p v-if="prop.description">{{prop.description}}</p>-->
-              <p v-html="prop.description"></p>
-            </li>
-          </ul>
-        </div>
-        <div class="fromMD">
-          <p v-if="param.description" v-html="param.description"></p>
-        </div>
-      </li>
-    </ul>
-  </div>
+  <func-params :func="func"></func-params>
   <template v-if="func.examples">
     <p class="arguments-title">Examples:</p>
     <example